fix input and output

+++ samples/pn_dev/app_data.h
@@ -57,7 +57,7 @@ extern "C" {
  *                           Not modified on error.
  * @return Reference to PNIO data, NULL on error
  */
-uint8_t * app_data_get_input_data (
+float * app_data_get_input_data (
     uint16_t slot_nbr,
     uint16_t subslot_nbr,
     uint32_t submodule_id,
